apps/api/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
python-multipart>=0.0.9,<1.0
PyYAML>=6.0,<7.0
python-dotenv>=1.0,<2.0
